requirements.txt
openai>=1.0.0
python-dotenv==1.0.0
PyPDF2==3.0.1
elevenlabs==0.2.27 